fix: add setTimeouts and remove listener
Also refactor server
